+++ src/model/items/AbstractNonMagicalItem.java
@@ -11,7 +11,7 @@ import model.units.*;
 
 public abstract class AbstractNonMagicalItem extends AbstractItem implements INonMagicalItem{
 
-    public AbstractNonMagicalItem(final String name, final int power, final int minRange, final int maxRange) {
+    public AbstractNonMagicalItem(final String name, final double power, final int minRange, final int maxRange) {
         super(name,power,minRange,maxRange);
     }
