+++ src/model/items/Axe.java
@@ -24,7 +24,7 @@ public class Axe extends AbstractNonMagicalItem {
    * @param maxRange
    *     the maximum range of the axe
    */
-  public Axe(final String name, final int power, final int minRange, final int maxRange) {
+  public Axe(final String name, final double power, final int minRange, final int maxRange) {
     super(name, power, minRange, maxRange);
   }
