+++ src/model/items/Darkness.java
@@ -27,7 +27,7 @@ public class Darkness extends AbstractMagicalItem{
      * @param maxRange
      *     the maximum range of the darkness
      */
-    public Darkness(final String name, final int power, final int minRange, final int maxRange) {
+    public Darkness(final String name, final double power, final int minRange, final int maxRange) {
         super(name, power, minRange, maxRange);
     }
